git-svn-id: svn+ssh://svn.code.sf.net/p/swistrack/code@259 18c3c8b4-d610-0410-8eed-94a590f67c64

+++ code-branch-thomas/core/Random.h
@@ -70,17 +70,17 @@ public:
 	}
 
 	//! Returns a float with a gaussian distribution.
-	float Normal() {
+	double Normal() {
 		return smRandomNormal->Normal();
 	}
 
 	//! Returns a float with a gaussian distribution with a given standard deviation and mean.
-	float Normal(float mean, float stddev) {
+	double Normal(float mean, float stddev) {
 		return smRandomNormal->Normal()*stddev+mean;
 	}
 
 	//! Returns a float with an exponential distribution.
-	float Exponential() {
+	double Exponential() {
 		return smRandomExponential->Exponential();
 	}
