+++ code-branch-thomas/core/RandomExponential.h
@@ -14,16 +14,16 @@ class RandomExponential {
 protected:
 	RandomMersenneTwister *mRandomMersenneTwister;
 	unsigned long ke[256];
-	float we[256], fe[256];
+	double we[256], fe[256];
 
-	float efix(unsigned long jz, unsigned long iz);
+	double efix(unsigned long jz, unsigned long iz);
 
 public:
 	//! Constructor.
 	RandomExponential(RandomMersenneTwister *rmt);
 
 	//! Returns a double with an exponential distribution.
-	float Exponential() {
+	double Exponential() {
 		unsigned long jz=mRandomMersenneTwister->randInt();
 		unsigned long iz=jz & 255;
 		return (jz<ke[iz]) ? jz*we[iz] : efix(jz, iz);
